added javadocs

+++ src/main/java/ims/stephenwongc482/controller/ModifyProductController.java
@@ -3,6 +3,8 @@ package ims.stephenwongc482.controller;
 import ims.stephenwongc482.model.Inventory;
 import ims.stephenwongc482.model.Part;
 import ims.stephenwongc482.model.Product;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
@@ -11,12 +13,16 @@ import javafx.scene.control.cell.PropertyValueFactory;
 
 import java.io.IOException;
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
+import static ims.stephenwongc482.controller.MainController.searchAlert;
 import static ims.stephenwongc482.controller.NavController.navigate;
-import static ims.stephenwongc482.model.Inventory.getAllProducts;
-import static ims.stephenwongc482.model.Inventory.getProductIdCount;
+import static ims.stephenwongc482.model.Inventory.*;
 
+/**
+ * ModifyProductController class is used to modify a product.
+ */
 public class ModifyProductController implements Initializable {
     public static Product productToModify = null;
     public TextField modifyProductIdInput;
@@ -41,6 +47,7 @@ public class ModifyProductController implements Initializable {
     public TableColumn assPartName;
     public TableColumn assPartStock;
     public TableColumn assPartPrice;
+    public TextField partSearchInput;
 
     private String name;
     private double price;
@@ -64,7 +71,6 @@ public class ModifyProductController implements Initializable {
     public static void setProductToModify(Product product) {
         productToModify = product;
     }
-
 
     /**
      * initalizes app
@@ -93,6 +99,49 @@ public class ModifyProductController implements Initializable {
         allPartPriceCol.setCellValueFactory(new PropertyValueFactory<>("price"));
     }
 
+    /**
+     * creates a list of parts that match the search criteria
+     *
+     * @param partialName - partial name
+     */
+    private ObservableList<Part> searchPartByName(String partialName) {
+        if (partialName == null || partialName.isEmpty()) {
+            return Inventory.getAllParts();
+        } else {
+            ObservableList<Part> searchResults = FXCollections.observableArrayList();
+            for (Part part : Inventory.getAllParts()) {
+                if (part.getName().contains(partialName)) {
+                    searchResults.add(part);
+                }
+            }
+            return searchResults;
+        }
+    }
+    /**
+     * handles search input on main screen and searches for part by name or id
+     *
+     * @param actionEvent - add product button on main screen
+     */
+    @FXML
+    void handleMainSearchPartSubmit(ActionEvent actionEvent) {
+        allPartTable.setItems(getAllParts());
+        String searchInputText = partSearchInput.getText();
+        try {
+            if (Integer.parseInt(searchInputText) > 0 && Integer.parseInt(searchInputText) < getAllProducts().size() + 1) {
+                allPartTable.getSelectionModel().select(Integer.parseInt(searchInputText)-1);
+                return;
+            } else {
+                allPartTable.getSelectionModel().clearSelection();
+                allPartTable.setItems(searchPartByName(searchInputText));
+            }
+        } catch (NumberFormatException e) {
+            allPartTable.getSelectionModel().clearSelection();
+            allPartTable.setItems(searchPartByName(searchInputText));
+        }
+        if (allPartTable.getItems().size() == 0) {
+            searchAlert("Part");
+        }
+    }
 
     @FXML
     void handleSaveBtn(ActionEvent actionEvent) throws IOException {
@@ -193,6 +242,14 @@ public class ModifyProductController implements Initializable {
      */
     @FXML
     void handleRemoveAssPart(ActionEvent actionEvent) {
+        Alert alertConf = new Alert(Alert.AlertType.CONFIRMATION);
+        alertConf.setTitle("Associated Part");
+        alertConf.setHeaderText("Remove");
+        alertConf.setContentText("Are you sure you want to remove this part?");
+        Optional<ButtonType> result = alertConf.showAndWait();
+        if (result.get() != ButtonType.OK) {
+            return;
+        }
         productToModify.deleteAssociatedPart((Part)assPartTable.getSelectionModel().getSelectedItem());
     }
